Form inGroups.

+++ database/migrations/2018_09_14_000008_create_records_table.php
@@ -17,7 +17,7 @@ class CreateRecordsTable extends Migration
             $table->engine = 'InnoDB';
             $table->increments('id');
             $table->integer('record_type_id')->unsigned();
-            $table->integer('user_id')->unsigned()->nullable();
+            $table->bigInteger('user_id')->unsigned()->nullable();
             $table->string('field_1_value', 100);
             $table->string('field_2_value', 100);
             $table->string('field_3_value', 100);
